+++ MyCreditManager/MyCreditManager/CreditManager.swift
@@ -7,49 +7,6 @@
 
 import Foundation
 
-enum InputType {
-    case menu
-    case text
-}
-
-enum FunctionType {
-    case delete
-    case update
-    case check
-}
-
-enum CreditManageError: Error {
-    case invalidInput(InputType)
-    case emptyStudents(FunctionType)
-    case existStudent
-    case notFoundStudent
-    case invalidGradeInput
-    case emptyCredits
-    
-    var errorDescription: String {
-        switch self {
-        case .invalidInput(.menu):
-            return "잘못된 입력입니다. 1~5 사이의 숫자 혹은 X를 입력해주세요."
-        case .invalidInput(.text):
-            return "잘못된 입력입니다. 다시 확인해 주세요."
-        case .emptyStudents(.delete):
-            return "삭제할 학생이 없습니다. 먼저 학생을 추가해 주세요."
-        case .emptyStudents(.update):
-            return "성적을 추가(변경)할 학생이 없습니다. 먼저 학생을 추가해 주세요."
-        case .emptyStudents(.check):
-            return "성적을 조회할 학생이 없습니다. 먼저 학생을 추가해 주세요."
-        case .existStudent:
-            return "학생의 데이터는 이미 존재하므로 추가되지 않습니다."
-        case .notFoundStudent:
-            return "학생을 찾을 수 없습니다."
-        case .invalidGradeInput:
-            return "올바르지 않은 성적이 입력되었습니다. 다시 확인해 주세요."
-        case .emptyCredits:
-            return "학생은 아직 입력된 성적이 없습니다. 성적을 추가해 주세요."
-        }
-    }
-}
-
 class CreditManager {
     private var students: [Student] = []
 
@@ -80,7 +37,7 @@ class CreditManager {
                 print("프로그램을 종료합니다...")
                 break menuLoop
             default:
-                print("잘못된 입력입니다. 1~5 사이의 숫자 혹은 X를 입력해주세요.")
+                print(CreditManageError.invalidInput(.menu).localizedDescription)
                 continue
             }
         }
@@ -92,7 +49,7 @@ class CreditManager {
         guard let name: String = getText() else { return }
 
         if students.contains(where: { $0.name == name }) {
-            print("\(name) 학생은 이미 존재하는 학생입니다. 추가하지 않습니다.")
+            print("\(name)", CreditManageError.existStudent.localizedDescription)
         } else {
             let student = Student(name: name)
             students.append(student)
@@ -102,7 +59,7 @@ class CreditManager {
 
     private func deleteStudent() {
         guard !students.isEmpty else {
-            print("삭제할 학생이 없습니다. 먼저 학생을 추가해 주세요.")
+            print(CreditManageError.emptyStudents(.deleteStudent).localizedDescription)
             return
         }
 
@@ -114,13 +71,13 @@ class CreditManager {
             students.remove(at: index)
             print("\(name) 학생을 삭제했습니다.")
         } else {
-            print("\(name) 학생을 찾을 수 없습니다.")
+            print("\(name)", CreditManageError.notFoundStudent.localizedDescription)
         }
     }
 
     private func updateCredit() {
         guard !students.isEmpty else {
-            print("성적을 추가할 학생이 없습니다. 먼저 학생을 추가해 주세요.")
+            print(CreditManageError.emptyStudents(.update).localizedDescription)
             return
         }
 
@@ -133,7 +90,7 @@ class CreditManager {
         let (name, subject, inputCredit) = (texts[0], texts[1], texts[2])
         
         guard let credit = Credit(rawValue: inputCredit) else {
-            print("성적이 잘못 입력되었습니다. 다시 확인해주세요.")
+            print(CreditManageError.invalidGradeInput.localizedDescription)
             return
         }
         
@@ -141,13 +98,13 @@ class CreditManager {
             students[index].credits[subject] = credit
             print("\(name) 학생의 \(subject) 과목이 \(inputCredit)으로 추가(변경)되었습니다.")
         } else {
-            print("\(name) 학생을 찾을 수 없습니다.")
+            print("\(name)", CreditManageError.notFoundStudent.localizedDescription)
         }
     }
 
     private func deleteCredit() {
         guard !students.isEmpty else {
-            print("성적을 추가할 학생이 없습니다. 먼저 학생을 추가해 주세요.")
+            print(CreditManageError.emptyStudents(.deleteCredit))
             return
         }
 
@@ -162,13 +119,13 @@ class CreditManager {
             students[index].credits[subject] = nil
             print("\(name) 학생의 \(subject) 과목의 성적이 삭제 되었습니다.")
         } else {
-            print("\(name) 학생을 찾을 수 없습니다.")
+            print("\(name)", CreditManageError.notFoundStudent.localizedDescription)
         }
     }
 
     private func checkScore() {
         guard !students.isEmpty else {
-            print("성적을 확인할 학생이 없습니다. 먼저 학생을 추가해 주세요.")
+            print(CreditManageError.emptyStudents(.deleteCredit))
             return
         }
 
@@ -178,7 +135,7 @@ class CreditManager {
         
         if let student = students.first(where: { $0.name == name}) {
             guard !student.credits.isEmpty else {
-                print("\(name) 학생은 아직 성적이 입력된 과목이 없습니다. 성적을 추가해 주세요.")
+                print("\(name)", CreditManageError.emptyCredits.localizedDescription)
                 return
             }
             
@@ -188,7 +145,7 @@ class CreditManager {
             
             print("평점: \(student.score)")
         } else {
-            print("\(name) 학생을 찾을 수 없습니다.")
+            print("\(name)", CreditManageError.notFoundStudent.localizedDescription)
         }
     }
 }
@@ -200,14 +157,14 @@ extension CreditManager {
         let word = input.trimmingCharacters(in: .whitespaces)
 
         guard !word.isEmpty else {
-            print("잘못된 입력입니다. 다시 확인해주세요.")
+            print(CreditManageError.invalidInput(.text).localizedDescription)
             return nil
         }
         
         let inputValues = word.components(separatedBy: " ")
         
         guard inputValues.count == count else {
-            print("잘못된 입력입니다. 다시 확인해주세요.")
+            print(CreditManageError.invalidInput(.text).localizedDescription)
             return nil
         }
